Added networking abstracting for MarsWeatherApi in order to seperate concerns
The mars weather Api accesses and data via network. However, we wanted to seperate which networking library it uses to make it independent. MarsApi defines its own Networking requirements in MarsWeatherApiNetworking which is called Inversion Of Control. This allows us to continue to create a Adapter for either Alamofire, NSURLSessions or any other networking service which increases the reuseability